Integer types changes int vs Integer

+++ dfd/dfdAlgorithm/src/fdiscovery/fastfds/runner/FastFDs.java
@@ -50,7 +50,7 @@ public class FastFDs extends Miner {
 			FastFDs fastFDRunner = new FastFDs(inputFileProcessor);
 			
 			fastFDRunner.run();
-			System.out.println(String.format("Dependencies: %d.", fastFDRunner.minimalDependencies.getCount()));
+			System.out.println(String.format("Dependencies: %d.", Integer.valueOf(fastFDRunner.minimalDependencies.getCount())));
 			long timeFindFDs = System.currentTimeMillis();
 			System.out.println("Total time:\t" + (timeFindFDs - timeStart)/1000 + "s");
 			System.out.println(fastFDRunner.getDependencies());
@@ -80,10 +80,10 @@ public class FastFDs extends Miner {
 			resultFile = cli.getOptionValue("result");
 		}
 		if (cli.hasOption("columns")) {
-			numberOfColumns = Integer.valueOf(cli.getOptionValue("columns"));
+			numberOfColumns = Integer.valueOf(cli.getOptionValue("columns")).intValue();
 		}
 		if (cli.hasOption("rows")) {
-			numberOfRows = Integer.valueOf(cli.getOptionValue("rows"));
+			numberOfRows = Integer.valueOf(cli.getOptionValue("rows")).intValue();
 		}
 		ColumnFiles columnFiles = new ColumnFiles(new File(columnFileDirectory), numberOfColumns, numberOfRows);
 		long timeStart = System.currentTimeMillis();
@@ -105,19 +105,19 @@ public class FastFDs extends Miner {
 		if (!inputFileName.isEmpty()) {
 			outputBuilder.append(String.format("%s\t", inputFileName));
 		}
-		outputBuilder.append(String.format("%d\t", this.numberOfRows));
-		outputBuilder.append(String.format("%d\t", this.numberOfColumns));
+		outputBuilder.append(String.format("%d\t", Integer.valueOf(this.numberOfRows)));
+		outputBuilder.append(String.format("%d\t", Integer.valueOf(this.numberOfColumns)));
 		outputBuilder.append(String.format("%s\t", timeString));
-		outputBuilder.append(String.format("%d\t", this.minimalDependencies.getCount()));
-		outputBuilder.append(String.format("%d\t", this.minimalDependencies.getCountForSizeLesserThan(2)));
-		outputBuilder.append(String.format("%d\t", this.minimalDependencies.getCountForSizeLesserThan(3)));
-		outputBuilder.append(String.format("%d\t", this.minimalDependencies.getCountForSizeLesserThan(4)));
-		outputBuilder.append(String.format("%d\t", this.minimalDependencies.getCountForSizeLesserThan(5)));
-		outputBuilder.append(String.format("%d\t", this.minimalDependencies.getCountForSizeLesserThan(6)));
-		outputBuilder.append(String.format("%d\t", this.minimalDependencies.getCountForSizeGreaterThan(5)));
-		outputBuilder.append(String.format("%d\t", 0));
-		outputBuilder.append(String.format("%d\t", 0));
-		outputBuilder.append(String.format("%d\n", Runtime.getRuntime().totalMemory()));
+		outputBuilder.append(String.format("%d\t", Integer.valueOf(this.minimalDependencies.getCount())));
+		outputBuilder.append(String.format("%d\t", Integer.valueOf(this.minimalDependencies.getCountForSizeLesserThan(2))));
+		outputBuilder.append(String.format("%d\t", Integer.valueOf(this.minimalDependencies.getCountForSizeLesserThan(3))));
+		outputBuilder.append(String.format("%d\t", Integer.valueOf(this.minimalDependencies.getCountForSizeLesserThan(4))));
+		outputBuilder.append(String.format("%d\t", Integer.valueOf(this.minimalDependencies.getCountForSizeLesserThan(5))));
+		outputBuilder.append(String.format("%d\t", Integer.valueOf(this.minimalDependencies.getCountForSizeLesserThan(6))));
+		outputBuilder.append(String.format("%d\t", Integer.valueOf(this.minimalDependencies.getCountForSizeGreaterThan(5))));
+		outputBuilder.append(String.format("%d\t", Integer.valueOf(0)));
+		outputBuilder.append(String.format("%d\t", Integer.valueOf(0)));
+		outputBuilder.append(String.format("%d\n", Long.valueOf(Runtime.getRuntime().totalMemory())));
 		outputBuilder.append(String.format("#Memory: %s\n", Miner.humanReadableByteCount(Runtime.getRuntime().totalMemory(), false)));
 
 		try {
@@ -174,7 +174,7 @@ public class FastFDs extends Miner {
 			if (orig.isEmpty()) {
 				ColumnCollection lhs = new ColumnCollection(this.numberOfColumns);
 				
-				for (Integer lhsIndex : lhs.setCopy(rhsIndex).complement().getSetBits()) {
+				for (int lhsIndex : lhs.setCopy(rhsIndex).complement().getSetBits()) {
 					this.minimalDependencies.addRHSColumn(lhs.setCopy(lhsIndex), rhsIndex);
 				}
 			} 
@@ -202,7 +202,7 @@ public class FastFDs extends Miner {
 		}
 
 		// RECURSIVE CASE
-		for (Integer remainingColumn : currentOrder.getOrderedColumns()) {
+		for (int remainingColumn : currentOrder.getOrderedColumns()) {
 			DifferenceSets nextDifferenceSets = uncovered.removeCovered(remainingColumn);
 			PartialOrder nextOrder = new PartialOrder(nextDifferenceSets, remainingColumn);
 			Path nextPath = (Path) currentPath.addColumn(remainingColumn);
